Revert "practice_9"
