+++ src/main/java/ua/edu/sumdu/j2se/dudynskyi/tasks/LinkedTaskList.java
@@ -95,6 +95,14 @@ public class LinkedTaskList extends AbstractTaskList implements Cloneable {
         return null;
     }
 
+    public Task[] toArray() {
+        Task[] result = new Task[taskAmount];
+        for (int i = 0; i < taskAmount; i++) {
+            result[i] = getTask(i);
+        }
+        return result;
+    }
+
     @Override
     public Stream<Task> getStream() {
         return Stream.of(toArray());
